# df.dfpde faster execution