--- a/Core/df.m
+++ b/Core/df.m
@@ -141,6 +141,15 @@ Jr = 0;
 % MaxStep = limit maximum time step size during integration
 options = odeset('MaxStep', par.MaxStepFactor*0.1*abs(par.tmax - par.t0), 'RelTol', par.RelTol, 'AbsTol', par.AbsTol, 'NonNegative', [1,1,1,0]);
 
+%% Prepare variables for solver
+
+% prepare search for x in x_ihalf
+x_ihalf_length = length(x_ihalf);
+sampling_step = round(sqrt(x_ihalf_length));
+x_ihalf_padded = [x_ihalf, nan(1,sampling_step-mod(x_ihalf_length,sampling_step))];
+x_ihalf_padded_mat = reshape(x_ihalf_padded,sampling_step,[]);
+x_ihalf_sampled = x_ihalf_padded_mat(1,:);
+
 %% Call solver
 % inputs with '@' are function handles to the subfunctions
 % below for the: equation, initial conditions, boundary conditions
@@ -151,10 +160,11 @@ u = pdepe(par.m,@dfpde,@dfic,@dfbc,x,t,options);
 % C = Time-dependence prefactor
 % F = Flux terms
 % S = Source terms
-    function [C,F,S] = dfpde(x,t,u,dudx)   
-        % Get position point
-        i = find(x_ihalf <= x);
-        i = i(end);
+function [C,F,S] = dfpde(x,t,u,dudx)   
+    % Get position point
+    sampled_i = find(x_ihalf_sampled <= x);
+    sampled_i = sampled_i(end);
+    i = (sampled_i-1)*sampling_step + find(x == x_ihalf_padded_mat(:,sampled_i));
         
         switch g1_fun_type
             case 'constant'