--- a/Core/df.m
+++ b/Core/df.m
@@ -231,9 +231,7 @@ function [C,F,S] = dfpde(x,t,u,dudx)
     else
         gxt2 = g2_fun(g2_fun_arg, t)*gx2(i);
     end
-        
-        g = gxt1 + gxt2;
-        
+
         %% Variables
         V = u(1); n = u(2); p = u(3); 
 
@@ -262,8 +260,8 @@ function [C,F,S] = dfpde(x,t,u,dudx)
             
         % Source terms
         S_potential = q_over_eppmax_epp0*(-n+p-a+c+NANDNaniNcat(i));
-        S_electron = g - radset*B(i)*((n*p)-(ni_squared(i))) - SRHset*(((n*p)-ni_squared(i))/((taun(i)*(p+pt(i)))+(taup(i)*(n+nt(i)))));
-        S_hole     = g - radset*B(i)*((n*p)-(ni_squared(i))) - SRHset*(((n*p)-ni_squared(i))/((taun(i)*(p+pt(i)))+(taup(i)*(n+nt(i)))));
+        S_electron = gxt1 + gxt2 - radset*B(i)*((n*p)-(ni_squared(i))) - SRHset*(((n*p)-ni_squared(i))/((taun(i)*(p+pt(i)))+(taup(i)*(n+nt(i)))));
+        S_hole     = gxt1 + gxt2 - radset*B(i)*((n*p)-(ni_squared(i))) - SRHset*(((n*p)-ni_squared(i))/((taun(i)*(p+pt(i)))+(taup(i)*(n+nt(i)))));
         S = [S_potential; S_electron; S_hole];
         
         if N_ionic_species == 1 || N_ionic_species == 2  % Condition for cation and anion terms