--- a/Core/df.m
+++ b/Core/df.m
@@ -242,13 +242,10 @@ function [C,F,S] = dfpde(x,t,u,dudx)
 
     if N_ionic_species
             c = u(4);           % Include cation variable
-            dcdx = dudx(4);
             a = Nani(i);
         if N_ionic_species_two
             c = u(4);           % Include cation variable
             a = u(5);           % Include anion variable
-            dcdx = dudx(4);
-            dadx = dudx(5);
         end
     else
         c = Ncat(i);
@@ -277,13 +274,13 @@ function [C,F,S] = dfpde(x,t,u,dudx)
 
         % F_cation: Flux term for cations
         % dudx(4) is dcdx, gradient of mobile cation concentration
-        F_cation = K_cation*mobseti*mucat(i)*(c*dVdx + kBT*dcdx*(1 + c/(DOScat(i)-c)));
+        F_cation = K_cation*mobseti*mucat(i)*(c*dVdx + kBT*dudx(4)*(1 + c/(DOScat(i)-c)));
 
         if N_ionic_species_two % Condition for anion terms
 
             % F_anion: Flux term for anions
             % dudx(5) is dadx, gradient of mobile anion concentration
-            F_anion = K_anion*mobseti*muani(i)*(a*-dVdx + kBT*dadx*(1 + a/(DOSani(i)-a)));
+            F_anion = K_anion*mobseti*muani(i)*(a*-dVdx + kBT*dudx(5)*(1 + a/(DOSani(i)-a)));
 
             % F: Flux terms
             F = [F_potential; F_electron; F_hole; F_cation; F_anion];