The `df.dfpde` function runs millions of times in each simulation (e.g. for running the example in the readme, including `equilibrate` and `doJV`, it runs 3.3 million times).
Currently it accounts for half of the total simulation time.
These changes tries to increase its efficiency without making the code too much ugly.
Even faster code could be obtained moving all the `if` checks outside of the `df.dfpde` function and writing various versions of it including 0, 1 or 2 ions. But this would mean much uglier code.
In the current set of changes the less nice part is the repetition of the `S_potential` expression.
And the slowest part of the function are the two `find` lines, I tried various alternatives (like ismembc2 and builtin _ismemberhelper) with no luck.
I am not super sure that all and each of the commits improve the speed, but must do.

Before the changes, the run time of the `equilibrate` example from the readme was 25.3 s (on my laptop) and after the changes is 17.9 s (further decreases to 17.7 s after applying also #38; further decreases to 13.8 s modifying pdepe for passing over the positional index, see [this question](https://uk.mathworks.com/matlabcentral/answers/518716-how-to-get-positional-index-from-pdepe-into-pde)).
Before the changes, the run time of the `doJV` example from the readme was 31.2 s and after the changes is 19.4 s (further decreases to 18.0 s after applying also #38; further decreases to 14.5 s modifying pdepe for passing over the positional index, see [this question](https://uk.mathworks.com/matlabcentral/answers/518716-how-to-get-positional-index-from-pdepe-into-pde)).
Using Matlab's profile function over the equilibrate and the doJV commands, the execution time of df.dfpde gets reduced to the half.