@ilario 2020-04-17
TL;DR: no improvement in further tests

I tested a couple more alternatives, in which the if checks are outside the `dfpde` function.
Instead of writing many different versions of `dfpde` function, I tried to generate it from a string and the string generated depending on the number of mobile ions and the generation dependence with time.
The problem is that `str2func` can generate only anonymous functions.

In this example:  https://github.com/ilario/Driftfusion/blob/dfpde3/Core/df.m I generated a function for returning F and S values and this function was used by `dfpde` function. The time needed for calling the new function compensates the removal of the if checks. In order to use only one function for generating both F and S I used the `deal` function, but the additional time needed by the `deal` function compensated the difference between calling two or one function.

In this other example: https://github.com/ilario/Driftfusion/blob/dfpde2/Core/df.m in order to avoid to call more than one function, the whole `dfpde` was generated using `str2func`. The problem is that the `i` index cannot be calculated inside of it, so the `pdepe.m` Matlab's function has to be modified to pass the index to the `dfpde` function as an argument.

Neither of the two modifications resulted in a considerable improvement of performances (over the code contained in this pull request) but they made the code much more dirty, so the current PR is still my favourite option.