@barnesgroupICL 2020-07-18
I accidentally accepted all of the commits to df.dfpde faster execution #39 when I was intending going through each commit individually. Now it the pull request has moved into ‘closed’ status. I’m afraid I can’t easily see how to reopen it… maybe it’s possible for you to do it?

In general I think we need to find a balance here between efficiency of the code and readability hence not accepting all the suggested changes without the appropriate scrutiny.