@ilario 2020-07-19
> Does this genuinely speed things up? Looks like you still have to use find(x_ihalf <=x) but then have an additional calculation to perform?

I suppose you refer to this:

https://github.com/ilario/Driftfusion/blob/1fdbf23a4e9c5db4ec52c66178102738b93fdec4/Core/df.m#L220-L223

The find is usually the most time consuming step, you can check that with `profile on; df(run_some_simulation); profile viewer`.

To perform 2 small find is a bit faster than a large find (I tested it when making the change, I can repeat the test).

What would be much faster would be to have the index directly from pdepe (see https://uk.mathworks.com/matlabcentral/answers/518716-how-to-get-positional-index-from-pdepe-into-pde ) but this would mean to distribute a modified copy of pdepe (which likely we cannot do due to Matlab's license?).

> This makes a lot of sense. Is it more efficient to use 'otherwise' in the switch rather than case 0? I think case 0 is probably safer for future expansion?

I suppose this refers to:

https://github.com/ilario/Driftfusion/blob/1fdbf23a4e9c5db4ec52c66178102738b93fdec4/Core/df.m#L197

and it has no influence on the performance, as it is out of the dfpde function, so if you think that having case 0 is better, that's perfect for me.

> Now it the pull request has moved into ‘closed’ status. I’m afraid I can’t easily see how to reopen it… maybe it’s possible for you to do it?

I'll make a new PR including also the aforementioned `case 0` change.